requirements.txt
pycrypto>=2.6.1
requests>=2.18.1
web.py>=0.37
